+++ src/mimic/src/orm/base/validator/string/iso.rs
@@ -15,7 +15,7 @@ impl ValidatorString for Iso6391 {
         let s = s.to_string();
 
         if s.len() != 2 || !s.chars().all(|c| c.is_ascii_lowercase()) {
-            Err("invalid ISO 3166-1 alpha-2 country code".to_string())
+            Err(format!("invalid ISO 3166-1 alpha-2 country code {s}"))
         } else {
             Ok(())
         }