+++ src/mimic/src/schema/node/arg.rs
@@ -63,13 +63,11 @@ pub enum ArgNumber {
     I32(i32),
     I64(i64),
     I128(i128),
-    Isize(isize),
     U8(u8),
     U16(u16),
     U32(u32),
     U64(u64),
     U128(u128),
-    Usize(usize),
 }
 
 impl ValidateNode for ArgNumber {}