+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 field a U8 and validate it between 10 and 100
 - went through the mimic_design/imp directory and forced everything into traits to be tidy
 - moved type validators into the Item struct, so they can be used on basically anything
+- isize and usize make no sense so have been removed
 
 ## [0.3.5]
 - Map and List are back, and the Map logic has been taken out of Newtype, it's much cleaner this way.